third_party/WebKit/LayoutTests/http/tests/htmlimports/import-slow-child.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
window.childLoaded = false;
function parentLoaded()
{
    shouldBeTrue('window.childLoaded');
    finishJSTest();
}

</script>
<link id="leaf" rel="import" href="resources/import-slow-child-parent.html" onload="parentLoaded()">
</head>
<body>
<script>
window.jsTestIsAsync = true;
</script>
</body>
</html>
