third_party/WebKit/LayoutTests/http/tests/htmlimports/import-slow-child-expected.txt

PASS window.childLoaded is true
PASS successfullyParsed is true

TEST COMPLETE
